fixed styling on post and created new post item components for better readibility and organization
